fix: :bug: Fix AddToLibraryButton-bug
Fix AddToLibraryButton-bug by making sure checkmark being displayed depends on key consisting of both libraryID and imdbID so that this is unique for all movies. Also Added green success-alert when movie is successfully added to either favorites or a library. Also duplicated Pagination-component on HomePage to be both above and below MovieContainerGrid.